Added test for list.

+++ include/block_list.h
@@ -19,8 +19,8 @@ void *block_get_start(Block *block);
 
 void list_initialize(BlockList *list, size_t (*get_sort_key)(Block *));
 
-size_t list_find_by_key(BlockList *list, size_t key);
-size_t list_find_by_key_or_higher(BlockList *list, size_t key);
+int list_find_by_key(BlockList *list, size_t key);
+int list_find_by_key_or_higher(BlockList *list, size_t key);
 
 Block list_get_block(BlockList *list, size_t index);
 void list_add_block(BlockList *list, Block block);
